whitespace fixes and camera driver renaming
