+++ artifacts/cdh_breakout_board_subsystem_sim/code.py
@@ -4,13 +4,13 @@ which may talk to CDH. This testing artifact runs on a custom breakout board.
 """
 
 import asyncio
-import inter_subsystem_rs485
+from inter_subsystem_rs485 import cdh_em_board_rs485_receiver_task
 
 async def gathered_task():
     """
     Task to run all other tasks concurrently.
     """
-    await asyncio.gather(inter_subsystem_rs485.cdh_em_board_rs485_receiver_task())
+    await asyncio.gather(cdh_em_board_rs485_receiver_task())
 
 if __name__ == "__main__":
     asyncio.run(gathered_task())
