+++ artifacts/cdh_em_board/code.py
@@ -3,14 +3,14 @@ Artifact to be run on the CDH Engineering Model Board
 """
 
 import asyncio
-import inter_subsystem_rs485
+from inter_subsystem_rs485 import cdh_em_board_rs485_send_task
 
 async def gathered_task():
     """
     Task to run all other tasks concurrently.
     """
     await asyncio.gather(
-        inter_subsystem_rs485.cdh_em_board_rs485_send_task()
+        cdh_em_board_rs485_send_task()
     )
