+++ artifacts/cdh_samd51_breakout_board/code.py
@@ -4,14 +4,14 @@ a custom SamD51 breakout board.
 """
 
 import asyncio
-import inter_subsystem_rs485
+from inter_subsystem_rs485 import samd51_breakout_receiver_task
 
 async def gathered_task():
     """
     Task to run all other tasks concurrently.
     """
     await asyncio.gather(
-        inter_subsystem_rs485.samd51_breakout_receiver_task()
+        samd51_breakout_receiver_task()
     )
