+++ artifacts/nucleo_file_system/code.py
@@ -3,16 +3,16 @@ Code to open a file and do UART communication on a nucleo board.
 """
 
 import asyncio
-import inter_subsystem_rs485
+from inter_subsystem_rs485 import nucleo_rs485_sender_task
 
 async def gathered_task():
     """Runs all tasks in parallel."""
     await asyncio.gather(
-        inter_subsystem_rs485.nucleo_rs485_sender_task()
+        nucleo_rs485_sender_task()
     )
 
 if __name__ == "__main__":
-    with open("/sus.txt", "a") as fp:
+    with open("/sus.txt", "a", encoding="utf-8") as fp:
         fp.write("Wow I can't believe it worked!")
     asyncio.run(gathered_task())
     print('CDH code.py has been run successfully.')
