+++ src/drivers/ov5642.py
